# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A CRUD server for analyzing financial portfolios comprising stocks, bonds, and ETFs.",
        "title": "Portfolio API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/users/{user_id}/portfolios": {
            "get": {
                "description": "Get all portfolios belonging to a user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "List user's portfolios",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.PortfolioListItem"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.Objective": {
            "type": "string",
            "enum": [
                "Aggressive Growth",
                "Growth",
                "Income Generation",
                "Capital Preservation",
                "Mixed Growth/Income"
            ],
            "x-enum-varnames": [
                "ObjectiveAggressiveGrowth",
                "ObjectiveGrowth",
                "ObjectiveIncomeGeneration",
                "ObjectiveCapitalPreservation",
                "ObjectiveMixedGrowthIncome"
            ]
        },
...
        "models.PortfolioListItem": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "objective": {
                    "$ref": "#/definitions/models.Objective"
                },
                "portfolio_type": {
                    "$ref": "#/definitions/models.PortfolioType"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "models.PortfolioType": {
            "type": "string",
            "enum": [
                "Ideal",
                "Active",
                "Historic"
            ],
            "x-enum-varnames": [
                "PortfolioTypeIdeal",
                "PortfolioTypeActive",
                "PortfolioTypeHistoric"
            ]
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}